# public/api-docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "\n\nHeader -> api_key :  \n\n",
    "version": "1.0.0",
    "title": "E-commerce"
  },
  "host": "127.0.0.1:8000",
  "basePath": "/api",
...
  "paths": {
...
    "/user/showUser": {
      "post": {
        "tags": [
          "User"
        ],
        "summary": "User Details",
        "operationId": "ShowUser",
        "produces": [
          "application/json"
        ],
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "description": "Enter bearer token",
            "required": true,
            "type": "string"
          }

        ],
        "responses": {
          "200": {
            "description": "Data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "APIKeyHeader": {
      "type": "apiKey",
      "in": "header",
      "name": "api-key"
    }
  },
...
  "schemes": [
    "http",
    "https"
  ]
}